+++ docs/whats-new.rst
@@ -13,6 +13,10 @@ Breaking changes
 Enhancements
 ~~~~~~~~~~~~
 
+- New function ``cfg.add_to_basenames`` now allows users to define their own
+  entries in glacier directories (:issue:`731`).
+  By `Fabien Maussion <https://github.com/fmaussion>`_.
+
 Bug fixes
 ~~~~~~~~~